refactor: remove redundant dataclasses and move to types module
