+++ derive_client/clients/__init__.py
@@ -1,7 +1,7 @@
 """Clients module"""
 
 from .async_client import AsyncClient
-from .base_client import BaseClient, ApiException
+from .base_client import ApiException, BaseClient
 from .http_client import HttpClient
 from .ws_client import WsClient
