--- derive_client/create_subaccount_module.py
@@ -1,37 +1,0 @@
-from dataclasses import dataclass
-
-from eth_abi.abi import encode
-from lyra_v2_action_signing.module_data import ModuleData
-from lyra_v2_action_signing.utils import decimal_to_big_int
-from web3 import Web3
-
-
-@dataclass
-class CreateSubAccountDetails:
-    amount: int
-    base_asset_address: str
-    sub_asset_address: str
-
-    def to_eth_tx_params(self):
-        return (
-            decimal_to_big_int(self.amount),
-            Web3.to_checksum_address(self.base_asset_address),
-            Web3.to_checksum_address(self.sub_asset_address),
-        )
-
-
-@dataclass
-class CreateSubAccountData(ModuleData):
-    amount: int
-    asset_name: str
-    margin_type: str
-    create_account_details: CreateSubAccountDetails
-
-    def to_abi_encoded(self):
-        return encode(
-            ['uint256', 'address', 'address'],
-            self.create_account_details.to_eth_tx_params(),
-        )
-
-    def to_json(self):
-        return {}
